UCS/SYS/TOPO: moving code changes to topo files
