+++ src/ucs/Makefile.am
@@ -101,6 +101,7 @@ noinst_HEADERS = \
 	sys/compiler.h \
 	sys/module.h \
 	sys/sys.h \
+	sys/topo.h \
 	sys/iovec.h \
 	time/time.h \
 	time/timerq.h \
@@ -154,6 +155,7 @@ libucs_la_SOURCES = \
 	sys/module.c \
 	sys/string.c \
 	sys/sys.c \
+	sys/topo.c \
 	sys/iovec.c \
 	sys/sock.c \
 	sys/stubs.c \
